requirements.txt
# 基础
numpy>=1.23
scipy>=1.10
pandas>=1.5
matplotlib>=3.7
opencv-python-headless>=4.8
scikit-image>=0.21
albumentations>=1.3
Pillow>=10.0

# 深度学习
torch>=2.0
torchvision>=0.15
timm>=0.9.2
pytorch-lightning>=2.3.0
# MONAI（可选）
monai>=1.3.0

# 配置/日志
hydra-core>=1.3
omegaconf>=2.3

# 评估
scikit-learn>=1.3

# Web 演示（可选）
streamlit>=1.32

# 代码质量（本地/CI）
black>=24.3
isort>=5.13
ruff>=0.5.0
mypy>=1.8

# 测试
pytest>=8.0
PyYAML~=6.0.1